Original change: https://android-review.googlesource.com/c/platform/packages/modules/Virtualization/+/1788638

+++ authfs/service/src/authfs.rs
@@ -54,7 +54,7 @@ impl Interface for AuthFs {}
 impl IAuthFs for AuthFs {
     fn openFile(
         &self,
-        remote_fd_name: i64,
+        remote_fd_name: i32,
         writable: bool,
     ) -> binder::Result<ParcelFileDescriptor> {
         let mut path = PathBuf::from(&self.mountpoint);
